Moved DatabaseSettings & LocalStorageSettings to own class.
