+++ src/main/java/me/wolfyscript/customcrafting/configs/MainConfig.java
@@ -177,106 +177,4 @@ public class MainConfig extends YamlConfiguration {
         return getBoolean("recipes.brewing");
     }
 
-    public static class LocalStorageSettings implements ConfigurationSerializable {
-
-        private static final String LOAD = "load";
-        private static final String BEFORE_DATABASE = "before_database";
-        private static final String OVERRIDE = "override";
-
-        private final boolean load;
-        private final boolean beforeDatabase;
-        private final boolean override;
-
-        public LocalStorageSettings(Map<String, Object> values) {
-            this.load = values.get(LOAD) instanceof Boolean bool && bool;
-            this.beforeDatabase = values.get(BEFORE_DATABASE) instanceof Boolean bool && bool;
-            this.override = values.get(OVERRIDE) instanceof Boolean bool && bool;
-        }
-
-        @Override
-        public @NotNull Map<String, Object> serialize() {
-            Map<String, Object> result = new HashMap<>();
-            result.put(LOAD, load);
-            result.put(BEFORE_DATABASE, beforeDatabase);
-            result.put(OVERRIDE, override);
-            return result;
-        }
-
-        public boolean isEnabled() {
-            return load;
-        }
-
-        public boolean isBeforeDatabase() {
-            return beforeDatabase;
-        }
-
-        public boolean isOverride() {
-            return override;
-        }
-    }
-
-
-    public static class DatabaseSettings implements ConfigurationSerializable {
-
-        private static final String ENABLED = "enabled";
-        private static final String HOST = "host";
-        private static final String PORT = "port";
-        private static final String SCHEMA = "schema";
-        private static final String USERNAME = "username";
-        private static final String PASSWORD = "password";
-
-        private final boolean enabled;
-        private final String host;
-        private final int port;
-        private final String schema;
-        private final String username;
-        private final String password;
-
-        public DatabaseSettings(Map<String, Object> values) {
-            this.enabled = values.get(ENABLED) instanceof Boolean bool && bool;
-            this.host = (String) values.getOrDefault(HOST, "localhost");
-            Object portVal = values.get(PORT);
-            this.port = portVal instanceof Number ? NumberConversions.toInt(portVal) : 3306;
-            this.schema = values.getOrDefault(SCHEMA, "mc_plugins").toString();
-            this.username = values.getOrDefault(USERNAME, "minecraft").toString();
-            this.password = values.getOrDefault(PASSWORD, "").toString();
-        }
-
-        @Override
-        public @NotNull Map<String, Object> serialize() {
-            Map<String, Object> result = new HashMap<>();
-            result.put(ENABLED, enabled);
-            result.put(HOST, host);
-            result.put(PORT, port);
-            result.put(SCHEMA, schema);
-            result.put(USERNAME, username);
-            result.put(PASSWORD, password);
-            return result;
-        }
-
-        public boolean isEnabled() {
-            return enabled;
-        }
-
-        public String getHost() {
-            return host;
-        }
-
-        public int getPort() {
-            return port;
-        }
-
-        public String getSchema() {
-            return schema;
-        }
-
-        public String getUsername() {
-            return username;
-        }
-
-        public String getPassword() {
-            return password;
-        }
-    }
-
 }
